backend/requirements.txt
fastapi[all]>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-dotenv>=1.0.0
geopandas>=0.14.0
pandas>=2.1.2
shapely>=2.0.2
requests>=2.31.0
# Added for caching
redis>=5.0.1
fastapi-cache2>=0.2.1
# Added for logging
python-json-logger>=2.0.7
loguru>=0.7.2